Improve handling of clang
Instead of blindly assuming clang doesn't support __builtin_va_arg_pack,
check for it with __has_builtin - clang 3.6 may add the builtin.

Don't #define snprintf if we're using C++ because some libraries and
applications (e.g. Chromium) try to pull snprintf into a namespace,
which fails if it's #defined. This is analogous to what glibc does with
the snprintf redefine.

Change-Id: I804e9a89e13d0d90ebca68d095b9e3f1677ae62b
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>